+++ src/Console/Application.php
@@ -120,6 +120,7 @@ EOF;
         $commands[] = new \SugarCli\Console\Command\Backup\DumpDatabaseCommand();
         $commands[] = new \SugarCli\Console\Command\Backup\DumpAllCommand();
         $commands[] = new \SugarCli\Console\Command\Backup\RestoreFilesCommand();
+        $commands[] = new \SugarCli\Console\Command\Backup\RestoreDatabaseCommand();
         $commands[] = new \SugarCli\Console\Command\CleanLangFilesCommand();
         $commands[] = new \SugarCli\Console\Command\Code\ButtonCommand();
         $commands[] = new \SugarCli\Console\Command\Code\ExecuteFileCommand();